Fixed score showing and player health after death

+++ Assets/Scripts/CORE/Gameplay/Score.cs
@@ -2,6 +2,6 @@ namespace CORE.Gameplay
 {
    public class Score
    {
-      public float ScoreValue { get; set; }
+      public int ScoreValue { get; set; }
    }
 }
